fix(backend): change type of arg 'deadline' to Float
change type of arg 'deadline' from Int to Float, once graphql doesn't support integers larger than 32 bits

+++ backend/src/graphql/modules/Projects/schema.ts
@@ -68,7 +68,7 @@ export class EditProjectArgs {
   @Field({ nullable: true })
   description?: string;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   deadline?: number;
 
   @Field(() => String, { nullable: true })
@@ -98,7 +98,7 @@ export class AddProjectArgs {
   @Field({ nullable: true })
   description?: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   deadline: number;
 
   @Field(() => String, { nullable: true })
